+++ bin/miden-cli/src/errors.rs
@@ -1,6 +1,6 @@
 use std::error::Error as StdError;
 
-use miden_client::ClientError;
+use miden_client::{authenticator::keystore::KeyStoreError, ClientError};
 use miden_objects::{AccountError, AccountIdError, AssetError};
 use miette::Diagnostic;
 use thiserror::Error;
@@ -45,6 +45,9 @@ pub enum CliError {
     #[error("io error")]
     #[diagnostic(code(cli::io_error))]
     IO(#[from] std::io::Error),
+    #[error("keystore error")]
+    #[diagnostic(code(cli::keystore_error))]
+    KeyStore(#[source] KeyStoreError),
     #[error("missing flag: {0}")]
     #[diagnostic(code(cli::config_error), help("Check the configuration file format."))]
     MissingFlag(String),