+++ crates/rust-client/src/mock.rs
@@ -29,6 +29,7 @@ use tonic::Response;
 use uuid::Uuid;
 
 use crate::{
+    authenticator::{keystore::FilesystemKeyStore, ClientAuthenticator},
     rpc::{
         domain::{
             account::{AccountDetails, AccountProofs},
@@ -42,7 +43,7 @@ use crate::{
         },
         NodeRpcClient, RpcError,
     },
-    store::{sqlite_store::SqliteStore, StoreAuthenticator},
+    store::sqlite_store::SqliteStore,
     transaction::ForeignAccount,
     Client,
 };
@@ -311,7 +312,7 @@ impl NodeRpcClient for MockRpcApi {
 // HELPERS
 // ================================================================================================
 
-pub async fn create_test_client() -> (MockClient, MockRpcApi) {
+pub async fn create_test_client() -> (MockClient, MockRpcApi, FilesystemKeyStore) {
     let store = SqliteStore::new(create_test_store_path()).await.unwrap();
     let store = Arc::new(store);
 
@@ -320,12 +321,14 @@ pub async fn create_test_client() -> (MockClient, MockRpcApi) {
 
     let rng = RpoRandomCoin::new(coin_seed.map(Felt::new));
 
-    let authenticator = StoreAuthenticator::new_with_rng(store.clone(), rng);
+    let keystore = FilesystemKeyStore::new(temp_dir()).unwrap();
+
+    let authenticator = ClientAuthenticator::new(rng, keystore.clone());
     let rpc_api = MockRpcApi::new();
     let boxed_rpc_api = Box::new(rpc_api.clone());
 
-    let client = MockClient::new(boxed_rpc_api, rng, store, Arc::new(authenticator), true);
-    (client, rpc_api)
+    let client = MockClient::new(boxed_rpc_api, rng, store, Arc::new(authenticator.clone()), true);
+    (client, rpc_api, keystore)
 }
 
 pub fn create_test_store_path() -> std::path::PathBuf {