+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 
 * Add check for empty pay to id notes (#714).
+* [BREAKING] Refactored authentication out of the `Client` and added new separate authenticators (#718).
 
 ## 0.7.0 (2025-01-28)